+++ src/Exception/RepositoryInformationException.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace CodeHqDk\RepositoryInformation\Exception;
+
+use Exception;
+
+class RepositoryInformationException extends Exception
+{
+}