+++ RELEASE_NOTES.md
@@ -3,4 +3,6 @@
 
 ## Version 0.0.0
 ### Changes
+* Adjust namespace
+* Added Exception
 * Added initial Models and Interfaces